requirements.txt
python-dotenv
Flask
langchain
openai==0.27.8
gunicorn==19.7.1
pymongo==4.4.0
dnspython==2.3.0
huggingface_hub
beautifulsoup4==4.11.2
lxml==4.9.2
httpx==0.23.3
requests==2.28.2
pinecone-client
langchain-pinecone
sentence-transformers
langchain-community>=0.0.38,<0.1.0
langchain-openai>=0.0.38,<0.1.0
feedparser